+++ src/main/java/com/team/mvc/database/services/CompanyService.java
@@ -25,7 +25,7 @@ public class CompanyService {
         companiesRepository.save(company);
     }
 
-    public boolean isCompanyNameUnique(Integer id, String companyName) {
+    public boolean isCompanyNameUnique(Long id, String companyName) {
         Companies company = companiesRepository.findByCompanyName(companyName);
         return (company == null || ((id != null) && (company.getCompanyId() == id)));
     }
